# Application connectivity validator performance improvements