--- a/components/central-application-connectivity-validator/cmd/centralapplicationconnectivityvalidator/centralapplicationconnectivityvalidator.go
+++ b/components/central-application-connectivity-validator/cmd/centralapplicationconnectivityvalidator/centralapplicationconnectivityvalidator.go
@@ -131,7 +131,7 @@ func main() {
 		log.WithContext().Error("Unable to start manager: %s", err.Error())
 		os.Exit(1)
 	}
-	if err = controller.NewController(log, mgr.GetClient(), idCache).SetupWithManager(mgr); err != nil {
+	if err = controller.NewController(log, mgr.GetClient(), idCache, options.appNamePlaceholder, options.eventingPathPrefixV1, options.eventingPathPrefixV2, options.eventingPathPrefixEvents).SetupWithManager(mgr); err != nil {
 		log.WithContext().Error("Unable to create reconciler: %s", err.Error())
 		os.Exit(1)
 	}