--- a/components/central-application-connectivity-validator/internal/controller/cachesynccontroller.go
+++ b/components/central-application-connectivity-validator/internal/controller/cachesynccontroller.go
@@ -31,6 +31,9 @@ func (c *controller) Reconcile(ctx context.Context, request reconcile.Request) (
 }
 
 func (c *controller) SetupWithManager(mgr ctrl.Manager) error {
+
+	c.cacheSync.Init(context.Background())
+
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&v1alpha1.Application{}).
 		Complete(c)