--- a/components/central-application-connectivity-validator/internal/validationproxy/handler.go
+++ b/components/central-application-connectivity-validator/internal/validationproxy/handler.go
@@ -250,7 +250,7 @@ func createReverseProxy(log *logger.Logger, destinationHost string, reqOpts ...r
 			}).DialContext,
 			MaxIdleConns:          400,
 			DisableKeepAlives:     false,
-			MaxIdleConnsPerHost:   40,
+			MaxIdleConnsPerHost:   2,
 			MaxConnsPerHost:       40,
 			ForceAttemptHTTP2:     false,
 			IdleConnTimeout:       10 * time.Second,