@Ressetkk 2023-07-14
/override pre-kyma-artifacts